+++ src/trice64.h
@@ -361,7 +361,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 //! \param v0 a 64 bit value
 #define Trice64m_1( tid, v0 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 8<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_1( v0 ) \
@@ -369,7 +369,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_2( tid, v0, v1 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 16<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_2( v0, v1); \
@@ -377,7 +377,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_3( tid, v0, v1, v2 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 24<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_3 ( v0, v1, v2 ); \
@@ -385,7 +385,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_4( tid, v0, v1, v2, v3 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 32<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_4( v0, v1, v2, v3 ); \
@@ -393,7 +393,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_5( tid, v0, v1, v2, v3, v4 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 40<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_5( v0, v1, v2, v3, v4 ); \
@@ -401,7 +401,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_6( tid, v0, v1, v2, v3, v4, v5 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 48<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_6( v0, v1, v2, v3, v4, v5 ); \
@@ -409,7 +409,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_7( tid, v0, v1, v2, v3, v4, v5, v6 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 56<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_7( v0, v1, v2, v3, v4, v5, v6 ); \
@@ -417,7 +417,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_8( tid, v0, v1, v2, v3, v4, v5, v6, v7 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 64<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_8( v0, v1, v2, v3, v4, v5, v6, v7 ); \
@@ -425,7 +425,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_9( tid, v0, v1, v2, v3, v4, v5, v6, v7, v8 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 72<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_9( v0, v1, v2, v3, v4, v5, v6, v7, v8 ); \
@@ -433,7 +433,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_10( tid, v0, v1, v2, v3, v4, v5, v6, v7, v8, v9 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 80<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_10( v0, v1, v2, v3, v4, v5, v6, v7, v8, v9 ); \
@@ -441,7 +441,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_11( tid, v0, v1, v2, v3, v4, v5, v6, v7, v8, v9, v10 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 88<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_11( v0, v1, v2, v3, v4, v5, v6, v7, v8, v9, v10 ); \
@@ -449,7 +449,7 @@ void trice64fn_12( uint16_t tid, uint64_t v0, uint64_t v1, uint64_t v2, uint64_t
 
 #define Trice64m_12( tid, v0, v1, v2, v3, v4, v5, v6, v7, v8, v9, v10, v11 ) \
     TRICE_ENTER \
-    uint64_t ts = TriceStamp16(); \
+    uint16_t ts = TriceStamp16(); \
     TRICE_PUT(0x80008000|(tid<<16)|tid); \
     TRICE_PUT( 96<<24 | (TRICE_CYCLE<<16) | ts ); \
     TRICE_PUT64_12( v0, v1, v2, v3, v4, v5, v6, v7, v8, v9, v10, v11 ) \
